+++ packages/portal-proto/src/features/cohortBuilder/CohortManager/CohortSelector.tsx
@@ -1,37 +1,60 @@
+import {
+  selectAvailableCohorts,
+  selectCurrentCohortId,
+  selectCurrentCohortModified,
+  selectCurrentCohortSaved,
+  setActiveCohort,
+  useCoreDispatch,
+  useCoreSelector,
+} from "@gff/core";
 import { Select } from "@mantine/core";
 import { CustomCohortSelectItem, UnsavedIcon } from "./CustomCohortSelectItem";
 import { FaCaretDown as DownArrowIcon } from "react-icons/fa";
+import { useDeepCompareCallback, useDeepCompareMemo } from "use-deep-compare";
 
-interface CohortSelectorProps {
-  cohortId: string;
-  cohortList: {
-    value: string;
-    label: string;
-    isSavedUnchanged: boolean;
-    cohortStatusMessage: string;
-  }[];
-  isSavedUnchanged: boolean;
-  cohortStatusMessage: string;
-  onCohortChange: (cohortId: string) => void;
-}
+const CohortSelector = () => {
+  const coreDispatch = useCoreDispatch();
+  const cohorts = useCoreSelector(selectAvailableCohorts);
+  const currentCohortId = useCoreSelector(selectCurrentCohortId);
+  const currentCohortModified = useCoreSelector(selectCurrentCohortModified);
+  const currentCohortSaved = useCoreSelector(selectCurrentCohortSaved);
+  const handleCohortChange = useDeepCompareCallback(
+    (id: string) => {
+      coreDispatch(setActiveCohort(id));
+    },
+    [coreDispatch],
+  );
 
-const CohortSelector: React.FC<CohortSelectorProps> = ({
-  cohortId,
-  cohortList,
-  isSavedUnchanged,
-  cohortStatusMessage,
-  onCohortChange,
-}: CohortSelectorProps) => {
+  const cohortStatusMessage = currentCohortSaved
+    ? "Changes not saved"
+    : "Cohort not saved";
+  const isSavedUnchanged = currentCohortSaved && !currentCohortModified;
+
+  const cohortList = useDeepCompareMemo(
+    () =>
+      cohorts
+        .sort((a, b) => (a.modified_datetime <= b.modified_datetime ? 1 : -1))
+        .map((x) => ({
+          value: x.id,
+          label: x.name,
+          isSavedUnchanged: x.saved && !x.modified,
+          cohortStatusMessage: x.saved
+            ? "Changes not saved"
+            : "Cohort not saved",
+        })),
+    [cohorts],
+  );
+
   return (
     <div className="flex flex-col pt-5" data-testid="cohort-list-dropdown">
       <Select
         data={cohortList}
         searchable
         clearable={false}
-        value={cohortId}
-        onChange={(cohortId) => {
-          if (cohortId !== null) {
-            onCohortChange(cohortId);
+        value={currentCohortId}
+        onChange={(id) => {
+          if (id !== null) {
+            handleCohortChange(id);
           }
         }}
         renderOption={CustomCohortSelectItem}
